XLCLOUD-1906 - Slurm runlists should be identical for all instances of the layer
All instances of the Slurm layer share identical Chef runlists.

The xlc-mco-agent cookbook is also executed before running the setup runlist because XLcloud custom
facts need to be available for the proper execution of the slurm cookbook.

Fix up the security groups on the Login instance too: do not use the default group and
allow all traffic to the XSA service.